go-kgp: Allow number of database managers to be configured
The default value is set to one, and I am uncertain if more are
required, but the infrastructure supports multiple goroutines to
process queries.

In addition to this, main now blocks on the organizer, not the
database manager. The interrupt signal handler must therefore close
the "queue" channel for the process to terminate.