go-kgp: Fix simple mode implementation
The server is not permitted to accept any move before the client has
sent a "yield" for every "stop" the server has sent. It is therefore
necessary to count the pending "yield" commands, and only accept a
"move" when the current pending count is equal to 0.

Negative values, ie. preemptive "yield" commands are treated as
illegal states and lead to an immediate termination of the connection.