go-kgp: Reimplement "organizer" as "queue"
To avoid synchronisation issues, three channels are used to
communicate with a "queue manager".  This operates on a local queue,
starting matches if possible.  This approach removes the need for
expliciy locking sychronisation and codition handling by having all
events being handled in a single thread.

The functions enqueue(*Client), promote(*Client) and forget(*Client)
were replaced by channels of the same name, resulting in no additional
complexity (the total line number count increased, but that is due to
mor extensive comments in the new implementation).

As a result, the interrupt handler was moved back to db.go, now
closing the database action channel.  When all the database managers
terminate, the database coordinator also terminates.  The main
function therefore reversts back to waiting for manageDatabase to
return, instead of organizer (now queueManager).