go-kgp: Distinguish between Game.Timeout and TCP.Timeout
The former is used to give the client a fixed amount of time to decide
on their move, while the latter is used over TCP to ensure the
connection is still alive.  It probably makes sense to have
TCP.Timeout < Game.Timeout, so that games with dead clients terminate
before the opponent tries to make a move.