go-kgp: Store choice in game object
This avoid issues with old game states being reused in new games,
leading to the possibility for illegal moves to be made.  Furthermore,
the g.ctrl had to be split into two channels for moving and yielding.
This change also obsoleted the need for the (weak) Action abstraction,
that was hence also removed due to overlapping changes.