go-kgp: Handle the death of an agent more gracefully
If the death signal coincides with the end of a game, then the
previous error handling might trigger an unintended panic, as the
other client might have been reassigned before the game thread
actually takes notice of this change.  As we are then certain that the
game has ended, we can end the thread immediatly.