go-kgp: Ensure an agent is not reassigned when the game is over
Related to the changes in a0fa8f4, this will check if game is
currently still responsible for the agent or not.  If not, the client
is not re-added to the waiting queue.  This change is not strictly
necessary, or completly correct, as there is the chance that an agent
is reassigned in the time between the check and until it is inserted
into the channel.  This should not be a problem though, as queue
checks for duplicates and removes these.