go-kgp: Fix off-by-one issue in match
As we are not iterating over queue, but queue[1:], the variable I
cannot be used to access the i'th element of queue.  Doing so results
in inconsistent modifications, and among other things reenqueueing an
agent to play a game even though it is actually playing (the
exact reason why this happens remains uncertain).  This change fixes
the assumption and has i be an index in queue, not queue[1:].