go-kgp: Use actual time instead of insertion time for moves
This prevents reordering of moves if when using multiple database
managers, a database action manages to overtake a previous request and
thus stores an illegal move in the database.  As the game is
reconstructed later on by ordering the moves by their timestamp, the
server might attempt an illegal move that would lead to an illegal
move, if Board.Sow wouldn't panic instead.