go-kgp: Double-check if a move has the right reference
In case the game has progressed to the next round /after/ the protocol
interpreter has checked if the reference is valid, the move will be
placed into the move channel.  When processing the next event it is
necessary to check if the move is still referencing the right state or
not, otherwise an illegal move might be triggered.