go-kgp: Simplify queue assumptions
By ensuring that there are no duplicates in the queue (as might be
necessary for a bot playing multiple games in freeplay mode), the more
complicated remove function can be replaced by a simple iteration.

This corresponds to the current assumptions around Game.Start, that
require the client to play only one game at a time.