jkgp: Send back a ping argument if given
This implements the same behaviour as kgp.py.  Note that this is not
yet standardised, but possible as it is outside the scope of the
defined behaviour (ping expects no arguments).

The intention here is to prevent the client from being able to
preemptively send "pong", without the server knowing if these are in
response to a ping or just anticipated.