go-kgp: Organise a round robin tournament out of order
As commented on in a previously (see diff), round robin does not
require each round to take place in some specific order.  By
listing all games that should take place beforehand, and then tracking
the available clients, we can simplify the implementation and
accelerate the tournament by avoiding the necessity to wait for the
slowest clients in each round.