go-kgp: Restart an isolated client after each game
To ensure fairness between clients in tournaments, everyone has to
start a new game with a fresh state, and no memory of previous games.
This means that the client ReadWriteCloser has to be exchanged with
the RWC from a new connection.

As this is tricky, this commit also includes mode debug logging, to
understand what is going on.