go-kgp: Buffer Game.death channel
This is necessary when a client died and the game has to be aborted
from outside, without blocking to prevent the scheduler lock from
being held too long or even provoking a deadlock.

Furthermore, if both clients die, one of the two shouldn't be waiting
for the death indicator to be handled by the now terminated game
manager (hence the buffer size of 2).