go-kgp: Update client entry in database more frequently
This is mainly intended for tournaments, where the token is set by the
tournament scheduler, so that the name and authors are still stored in
the database, even if the token is ignored.

The issue here is that there is more database action going on, which
should be considered and evaluated again at some later point.