[go-kgp] Maintain a seperate read and write connection to the database
It is necessary to restrict the number of open connections to the
database to one (which is done using .SetMaxOpenConns), but this
proves to be a significant bottleneck.  This is not an issue when
writing to the database, as these operations are expected to be
blocking.  The main issues is had with the web interface, that appears
to suffer when there is a high write-load.

What we do instead is to provide two connections, one for writing and
one for reading.  The SQL files starting with "select-" are prepared
using "read", the remaining ones using "write".  The "write"
connection is initialised with .SetMaxOpenConns(1) resolving locking
issues that may arise when two threads attempt to modify conflicting
data, especially when at least one of the two is currently in a
transaction.

Due to the use of WAL, "write" does not prevent "read" from serving
information to the web interface.