[go-kgp] Simplify locking while regenerating the dominance graph
The previous could run into an issue if gen() ran for more than one
minute, as that would allow give new requests an opportunity to
increment the counter, at which point we have a data-race in who will
write their results into "data" first.

The new approach requires a second write-back to indicate that the
next request, if "time.Now().After(next)" has evaluated to true may
proceed to re-generate the graph.