[go-kgp] Ensure dead clients are not scheduled
We require an agent states if it is alive or not.  If an agent is
dead, it will be removed from the queue, which prevents
pseudo-resignations after the first move.

Note that it is not possible to make this a stricter condition by
panicing if a game starts with a dead client, as it is possible for a
client to die between the check and the actual beginning of a game.