[go-kgp] Do not reset the timeout timer on every response
This is embarrassing.

Interestingly enough it works for well-behaved and stable algorithms
that converge on a single move, e.g. when using iterated depth search.
As the search depth increases, they take more and more time to
respond, and eventually the timeout interrupts them even though they
had been given more time than was intended.  This breaks if and agent
fails to converge during a later stage of the game and is manages to
calculating oscillating decisions faster than the move timeout,
thereby resetting the timer over and over again.